requirements.txt
google-ads>=29.0.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
requests>=2.32.3
python-dateutil>=2.9.0
